Major Refactor
This is a package, damn it!
