+++ LootTools/__init__.py
@@ -1,0 +1,1 @@
+from .loottracker import LootTracker, LootBag
